+++ src/xataWorkspace/xataWorkspaceResponses.ts
@@ -5,32 +5,37 @@
  */
 import type * as Schemas from "./xataWorkspaceSchemas";
 
-export type SimpleError = {
+export type BadRequestError = {
   id?: string;
   message: string;
 };
 
-export type BulkError = {
-  errors: {
-    message?: string;
-    status?: number;
-  }[];
+/**
+ * @example {"message":"invalid API key"}
+ */
+export type AuthError = {
+  id?: string;
+  message: string;
 };
 
-export type BulkInsertResponse =
-  | {
-      recordIDs: string[];
-    }
-  | {
-      records: Schemas.XataRecord[];
-    };
+export type SimpleError = {
+  id?: string;
+  message: string;
+};
 
 export type BranchMigrationPlan = {
   version: number;
   migration: Schemas.BranchMigration;
 };
 
-export type RecordResponse = Schemas.XataRecord;
+export type SchemaUpdateResponse = {
+  /*
+   * @minLength 1
+   */
+  migrationID: string;
+  parentMigrationID: string;
+  status: Schemas.MigrationStatus;
+};
 
 export type SchemaCompareResponse = {
   source: Schemas.Schema;
@@ -47,31 +52,28 @@ export type RecordUpdateResponse =
       };
     };
 
+export type PutFileResponse = Schemas.FileResponse;
+
+export type RecordResponse = Schemas.XataRecord;
+
+export type BulkInsertResponse =
+  | {
+      recordIDs: string[];
+    }
+  | {
+      records: Schemas.XataRecord[];
+    };
+
+export type BulkError = {
+  errors: {
+    message?: string;
+    status?: number;
+  }[];
+};
+
 export type QueryResponse = {
   records: Schemas.XataRecord[];
   meta: Schemas.RecordsMetadata;
-};
-
-export type SchemaUpdateResponse = {
-  /*
-   * @minLength 1
-   */
-  migrationID: string;
-  parentMigrationID: string;
-  status: Schemas.MigrationStatus;
-};
-
-export type SummarizeResponse = {
-  summaries: Record<string, any>[];
-};
-
-/**
- * @example {"aggs":{"dailyUniqueUsers":{"values":[{"key":"2022-02-22T22:22:22Z","uniqueUsers":134},{"key":"2022-02-23T22:22:22Z","uniqueUsers":90}]}}}
- */
-export type AggResponse = {
-  aggs?: {
-    [key: string]: Schemas.AggResponse;
-  };
 };
 
 export type SearchResponse = {
@@ -79,39 +81,25 @@ export type SearchResponse = {
   warning?: string;
 };
 
-/**
- * @x-go-type TxResponse
- */
-export type TransactionSucceeded = {
-  /*
-   * An ordered array of results from the submitted operations that were executed
-   */
-  results: (
-    | Schemas.TransactionResultInsert
-    | Schemas.TransactionResultUpdate
-    | Schemas.TransactionResultDelete
-  )[];
+export type SQLResponse = {
+  records: Schemas.SQLRecord[];
+  warning?: string;
 };
 
-/**
- * @x-go-type TxResponse
- */
-export type TransactionFailed = {
-  /*
-   * An array of errors from the submitted operations.
-   */
-  errors: Schemas.TransactionError[];
-};
-
-export type BadRequestError = {
+export type RateLimitError = {
   id?: string;
   message: string;
 };
 
+export type SummarizeResponse = {
+  summaries: Record<string, any>[];
+};
+
 /**
- * @example {"message":"invalid API key"}
+ * @example {"aggs":{"dailyUniqueUsers":{"values":[{"$count":321,"$key":"2022-02-22T22:22:22Z","uniqueUsers":134},{"$count":202,"$key":"2022-02-23T22:22:22Z","uniqueUsers":90}]}}}
  */
-export type AuthError = {
-  id?: string;
-  message: string;
+export type AggResponse = {
+  aggs?: {
+    [key: string]: Schemas.AggResponse;
+  };
 };